페이징 시작

+++ src/main/kotlin/SayingApp/repository/SayingRepository.kt
@@ -13,11 +13,12 @@ object SayingRepository {
         val saveSaying = Saying(++id, author, saying)
         sayingMap[id] = saveSaying
         saveToFile(saveSaying, id.toString())
+        saveToLastId(id.toString())
         return id
     }
     fun delete(id:Int):Int{
         sayingMap.remove(id)
-        deleteFile(id.toString())
+        deleteFromFile(id.toString())
         return id
     }
     fun modify(id:Int, author:String, saying:String):Int{
@@ -32,14 +33,28 @@ object SayingRepository {
     fun check(id:Int):Boolean{
         return sayingMap.containsKey(id)
     }
+    fun build(){
+        val json = Json.encodeToJsonElement(findAll())
+        File("db/build").writeText(json.toString())
+    }
+    fun search(term:String):List<Saying>{
+        val findList = mutableListOf<Saying>()
+        for(i in findAll()){
+            if (i.saying.contains(term)|| i.author.contains(term)){
+                findList.add(i)
+            }
+        }
+        return findList
+    }
     private fun saveToFile(saying: Saying, fileName: String){
         val json = Json.encodeToJsonElement(saying)
-        File(fileName).writeText(json.toString())
+        File("db/$fileName").writeText(json.toString())
     }
-    private fun deleteFile(fileName: String) {
-        val file = File(fileName)
-        if (file.exists()) {
-            file.delete()
-        }
+    private fun saveToLastId(id:String){
+        File("db/lastId").writeText(id)
+    }
+    private fun deleteFromFile(fileName: String) {
+        val file = File("db/$fileName")
+        if (file.exists()) file.delete()
     }
 }
